feature: 属性变量方式时增加数据示例
